Console output formatter improvement
Current SymfonyStyle can cause some problems:
- can cut format tags,
- cut/split URLs, the user can't click on them in the terminal

I fixed the problem and added or modified some tests. This is a simple,
fast, and working solution with a regular expression that won't wrap in
the middle of a formatting tag or a link. The last one is optional,
see the doc change.

Co-authored-by: Théo FIDRY <theo.fidry@gmail.com>